+++ CHANGES.md
@@ -5,6 +5,9 @@
   quite a bit of overhead.
   https://github.com/joblib/threadpoolctl/pull/250
 
+- Added official support for Python 3.15, and declared support for Python 3.14
+  in package metadata.
+  https://github.com/joblib/threadpoolctl/issues/251
 
 3.7.0 (2026-09-15)
 ==================